fix(app-shell): make coerce values a react hook, bail out if value is not a string
